+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * provider/resources/objectgroup.go: `range` related filters now have support, although currently disabled
 * provider/resources/objectgroup.go: `live_events` are now enabled
 * provider/resources/objectgroup.go: `compression` type is now specifiable in `options`
+* provider/resources/destinations.go: `destinations` are now a supported resource for kibana alerts
+* provider/resources/monitors.go: `monitors` are now a supported resource for kibana alerts
 
 ### Enhancements:
 * provider/resources/objectgroup.go: Pull apart `filter` to more reflect API. Extends supported `filter` types